+++ src/app/app.component.js
@@ -1,6 +1,8 @@
 import template from './app.html';
 import './app.scss';
 import changelog from '../views/changelog-toaster-release.html';
+import { isPhonePortrait, isPhonePortraitStream } from './mediaQueries';
+import { subscribeOnScope } from './rx-utils';
 import _ from 'underscore';
 import i18next from 'i18next';
 
@@ -37,6 +39,22 @@ function AppComponentCtrl(
     });
     $scope.$watch(() => this.settings.vaultMaxCol, (cols) => {
       document.querySelector('html').style.setProperty("--vault-max-columns", cols);
+    });
+
+    $scope.$watch(() => this.settings.charColMobile, (cols) => {
+      // this check is needed so on start up/load this doesn't override the value set above on "normal" mode.
+      if (isPhonePortrait()) {
+        document.querySelector('html').style.setProperty("--character-columns", cols);
+      }
+    });
+    // a subscribe on isPhonePortraitStream is needed when the user on mobile changes from portrait to landscape
+    // or a user on desktop shrinks the browser window below isphoneportrait treshold value
+    subscribeOnScope($scope, isPhonePortraitStream(), (isPhonePortrait) => {
+      if (isPhonePortrait) {
+        document.querySelector('html').style.setProperty("--character-columns", this.settings.charColMobile);
+      } else {
+        document.querySelector('html').style.setProperty("--character-columns", this.settings.charCol);
+      }
     });
 
     hotkeys = hotkeys.bindTo($scope);